// apps/admin-web/src/.generated/api/model/createEventRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */
import type { CreateEventRequestDtoImageUrl } from './createEventRequestDtoImageUrl';
import type { CreateEventRequestDtoLinkUrl } from './createEventRequestDtoLinkUrl';
import type { CreateEventRequestDtoStatus } from './createEventRequestDtoStatus';

export interface CreateEventRequestDto {
  /** @maxLength 255 */
  title: string;
  content: string;
  startsAt: string;
  endsAt: string;
  /** @maxLength 500 */
  imageUrl?: CreateEventRequestDtoImageUrl;
  /** @maxLength 500 */
  linkUrl?: CreateEventRequestDtoLinkUrl;
  status: CreateEventRequestDtoStatus;
}