// apps/admin-web/src/.generated/api/model/eventItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */
import type { EventItemDtoImageUrl } from './eventItemDtoImageUrl';
import type { EventItemDtoLinkUrl } from './eventItemDtoLinkUrl';
import type { EventItemDtoPublishedAt } from './eventItemDtoPublishedAt';
import type { EventItemDtoStatus } from './eventItemDtoStatus';

export interface EventItemDto {
  id: string;
  title: string;
  content: string;
  startsAt: string;
  endsAt: string;
  /** @nullable */
  imageUrl: EventItemDtoImageUrl;
  /** @nullable */
  linkUrl: EventItemDtoLinkUrl;
  status: EventItemDtoStatus;
  /** @nullable */
  publishedAt: EventItemDtoPublishedAt;
  createdAt: string;
  updatedAt: string;
}